requirements.txt
pathway==0.7.0
chromadb==0.4.24
sentence-transformers==2.2.2
pandas==2.1.3
openpyxl==3.11.0
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
langchain==0.1.1
ollama==0.1.15
regex==2023.12.25
pathlib-meta==1.0.0
